project/requirements.txt
fastapi==0.60.1
pandas==1.1.0
plotly==4.9.0
uvicorn==0.11.8
psycopg2-binary==2.8.4
python-dotenv==0.14.0
Jinja2==2.11.1
aiofiles==0.5.0